+++ test/FieldsetTest.php
@@ -17,6 +17,9 @@ use PHPUnit\Framework\TestCase;
 
 class FieldsetTest extends TestCase
 {
+    /** @var Fieldset */
+    private $fieldset;
+
     public function setUp()
     {
         $this->fieldset = new Fieldset();
@@ -606,5 +609,29 @@ class FieldsetTest extends TestCase
         $fieldset->setMessages($messages);
 
         $this->assertEquals($messages, $fieldset->getMessages());
+    }
+
+    public function testSetNullValueWhenArrayProvided()
+    {
+        $subValue = 'sub-element-value';
+        $subElement = new Element('subElement');
+        $this->fieldset->add($subElement);
+        $this->fieldset->populateValues(['subElement' => $subValue]);
+        $this->assertSame($subValue, $subElement->getValue());
+
+        $this->fieldset->populateValues(['subElement' => null]);
+        $this->assertNull($subElement->getValue());
+    }
+
+    public function testSetNullValueWhenTraversableProvided()
+    {
+        $subValue = 'sub-element-value';
+        $subElement = new Element('subElement');
+        $this->fieldset->add($subElement);
+        $this->fieldset->populateValues(new TestAsset\CustomTraversable(['subElement' => $subValue]));
+        $this->assertSame($subValue, $subElement->getValue());
+
+        $this->fieldset->populateValues(new TestAsset\CustomTraversable(['subElement' => null]));
+        $this->assertNull($subElement->getValue());
     }
 }